Fix new Process.spawn spec

+++ core/process/spawn_spec.rb
@@ -393,32 +393,43 @@ describe "Process.spawn" do
 
   # chdir
 
-  platform_is_not :windows do
-    context "chdir" do
+  platform_is :linux do
+    describe "inside Dir.chdir" do
       def child_pids(pid)
-        `pgrep -P #{pid}`.each_line.map { |p| p.strip.to_i }
+        `pgrep -P #{pid}`.lines.map { |pid| Integer(pid) }
       end
 
       it "does not create extra process without chdir" do
-        child_pids(Process.spawn("cat")).size.should == 0
+        pid = Process.spawn("sleep 10")
+        begin
+          child_pids(pid).size.should == 0
+        ensure
+          Process.kill("TERM", pid)
+          Process.wait(pid)
+        end
       end
 
       it "kills extra chdir processes" do
-        pid = Dir.chdir("/tmp") { Process.spawn("cat") }
+        pid = nil
+        Dir.chdir("/tmp") do
+          pid = Process.spawn("sleep 10")
+        end
 
         children = child_pids(pid)
-        children.size.should == 1
+        children.size.should <= 1
 
         Process.kill("TERM", pid)
-        Process.wait(pid, Process::WNOHANG)
+        Process.wait(pid)
 
-        # wait a bit for children to die
-        sleep(1)
+        if children.size > 0
+          # wait a bit for children to die
+          sleep(1)
 
-        children.each do |child|
-          lambda do
-            Process.kill("TERM", child)
-          end.should raise_error(Errno::ESRCH)
+          children.each do |child|
+            lambda do
+              Process.kill("TERM", child)
+            end.should raise_error(Errno::ESRCH)
+          end
         end
       end
     end
